# docs/api/openapi.yml
openapi: 3.0.0
info:
  title: Defcon API
  description: Defcon control plane API documentation
  version: 0.1.0
...
paths:
...
  /api/alerters:
    get:
      summary: List all alerters
      tags: ['Alerters']
      responses:
        '200':
          description: List of alerters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/alerter/ReadAlerter'

    post:
      summary: Create an alerter
      tags: ['Alerters']
      requestBody:
        description: The definition of an alerter
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/alerter/WriteAlerter'
      responses:
        '201':
          description: The alerter was created
          headers:
            Location:
              description: The absolute path to the alerter, including the unique identifier
              schema:
                type: string
                format: url
        '400':
          description: Invalid payload
  /api/alerters/{uuid}:
    get:
      summary: Get an alerter
      tags: ['Alerters']
      parameters:
        - name: uuid
          description: Unique identifier
          in: path
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The definition of the alerter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/alerter/ReadAlerter'
        '404':
          description: Provided UUID does not match a known alerter

    put:
      summary: Update an alerter
      tags: ['Alerters']
      parameters:
        - name: uuid
          description: Unique identifier
          in: path
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        description: The definition of an alerter
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/alerter/WriteAlerter'
      responses:
        '200':
          description: The alerter was updated
        '400':
          description: Invalid payload
        '404':
          description: Provided UUID does not match a known alerter

    patch:
      summary: Update alerter attributes
      tags: ['Alerters']
      parameters:
        - name: uuid
          description: Unique identifier
          in: path
          schema:
            type: string
            format: uuid
          required: true
      requestBody:
        description: The definition of an alerter
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/alerter/ReadAlerter'
      responses:
        '200':
          description: The alerter was updated
        '400':
          description: Invalid payload
        '404':
          description: Provided UUID does not match a known alerter
...